+++ src/ts/Core/Utilities/Localization.ts
@@ -11,6 +11,7 @@ import EnglishConsent from 'json/locale/en/consent.json';
 import SpanishEndscreen from 'json/locale/es/endscreen.json';
 import SpanishLoadingScreen from 'json/locale/es/loadingscreen.json';
 import SpanishOverlay from 'json/locale/es/overlay.json';
+import SpanishConsent from 'json/locale/es/consent.json';
 import FinnishEndscreen from 'json/locale/fi/endscreen.json';
 import FinnishLoadingScreen from 'json/locale/fi/loadingscreen.json';
 import FinnishOverlay from 'json/locale/fi/overlay.json';
@@ -39,6 +40,7 @@ import RomanianOverlay from 'json/locale/ro/overlay.json';
 import RussianEndscreen from 'json/locale/ru/endscreen.json';
 import RussianLoadingScreen from 'json/locale/ru/loadingscreen.json';
 import RussianOverlay from 'json/locale/ru/overlay.json';
+import RussianConsent from 'json/locale/ru/consent.json';
 import TurkishEndscreen from 'json/locale/tr/endscreen.json';
 import TurkishOverlay from 'json/locale/tr/overlay.json';
 import ChineseSimplifiedEndscreen from 'json/locale/zh_Hans/endscreen.json';
@@ -115,7 +117,8 @@ export class Localization {
         'ru.*': {
             'endscreen': JSON.parse(RussianEndscreen),
             'overlay': JSON.parse(RussianOverlay),
-            'loadingscreen': JSON.parse(RussianLoadingScreen)
+            'loadingscreen': JSON.parse(RussianLoadingScreen),
+            'consent': JSON.parse(RussianConsent)
         },
         'ja.*': {
             'endscreen': JSON.parse(JapaneseEndscreen),
@@ -154,7 +157,9 @@ export class Localization {
         'es.*': {
             'endscreen': JSON.parse(SpanishEndscreen),
             'overlay': JSON.parse(SpanishOverlay),
-            'loadingscreen': JSON.parse(SpanishLoadingScreen)
+            'loadingscreen': JSON.parse(SpanishLoadingScreen),
+            'consent': JSON.parse(SpanishConsent)
+
         },
         'fr.*': {
             'endscreen': JSON.parse(FrenchEndscreen),
@@ -208,7 +213,7 @@ export class Localization {
     private _namespace: string;
 
     constructor(language: string, namespace: string) {
-        if (namespace === 'consent' && !language.match('fr.*') && !language.match('de.*')) {
+        if (namespace === 'consent' && !this.isConsentTranslationAvailable(language)) {
             language = 'en.*';
         }
 
@@ -236,6 +241,17 @@ export class Localization {
             return (Math.floor(number / 1000)).toString() + ' ' + localizedAbbreviations.thousand;
         }
         return number.toString();
+    }
+
+    private isConsentTranslationAvailable(language: string) {
+        if (language.match('fr.*')
+            || language.match('de.*')
+            || language.match('es.*')
+            || language.match('ru.*')) {
+            return true;
+        }
+
+        return false;
     }
 
 }